requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyGithub>=2.1.1  # Only used by scripts/populate_metadata.py utility
cisco-ai-skill-scanner>=0.1.0
